Created scenarios and methods to test the new Home Secretary symbol visible in workstacks due to Data Input or Markup stage actions.
Refactored all DCU pages to use the helper methods, thereby allowing us to use the standard accordion check method in all workflow scenarios, increasing out regression coverage to case data as well as workflow data.

Reworked / reworded DCU scenarios so that they are less explicit and more implicit in their level of detail, as per best practise for BDD scenarios.

Added workflow scenario to each DCU feature for happy path, as we previously only had the endToEnd scenario. Unsure on how this will effect runtime of regression, and may have to find way to combine for efficiency in future.

Where appropriate, added assertions to DCU scenarios to test team ownership, user ownership, timeline note additions, and case data, extending the regression coverage considerable for DCU workflows.

Created DCUProgressCase file, which has all full stage completion methods for happy path of DCU workflows. This should make the endToEnd stepdefs a little more traceable/readable in future.

Removed DCUCaseDetailsAccordion feature and stepdefs, as I included the accordion check into each workflow scenario in DCU features.

Renamed AddCorrespondents to Correspondents, as the page contains methods for more than just adding the correspondents.

Reworked correspondents methods so that you can now select 'member' as one of the types of correspondents to add without specifying a named person. This means we arent dependant on specific entries in the list for most of the scenarios. Renamed the method for adding a named member so its use is clear.

Got rid of methods in DCU pages that were collections of page actions that constitute the scripts that belong in the stepdef file. This is an improvement that comes from better understanding of best practise since DCU scenarios were written, but will have no real effect at runtime, just in readability.

Refactored many DCU tenant specific steps from the BaseStepDef page to the relevant DCU specific stepdefs.

Refactored some timeline specific steps from the BaseStepDef page to the timelineStepDef page.

Deleted unused methods / stepdefs / imports where ever I found the. I did not delete unused WebElementFacades, as they could always come in handy in future.

Removed unnecessary switch cases from stepdefs where we are only using one case for any scenario, and no new scenarios to cover those functional areas are required.

Added new assertions for checking DCU drafting team / private office team ownership, as hard coding the team names did not make sense in the BDD scenario steps, as we had not made it clear in the background steps that these teams had been previously selected.

Removed all un-parameterised 'I complete the x stage' steps/methods for DCU stages, as we had both that and the parameterised version, and maintaining both made little sense.

Refactored Markup_AddTopic into Markup page, as they were originally separate as we thought more DECS workflows would use adding topics, but this is the only DECS stage that uses topics in this way.

Renamed page elements 'autoAssignedDraftTeam' and 'autoAssignedPrivateOfficeTeam' to 'defaultDraftTeam' and 'defaultPrivateOfficeTeam', as this is more accepted nomenclature.

Added Basepage methods for checking the current cases casetype against all existing casetypes (returns boolean), and used them in methods where we currently create a string from the casetype session variable.

Combined the stage check assertion step for FOI cases together with the one for all other casetypes, with if statement for FOI casetype.

Deleted the validation scenario for creating case notes from each DCU stage feature, as there is no functional difference to the timeline at each stage, and the scenario in the timeline feature is sufficient coverage for this functionality.

Reversed the parameters in the assertSummaryContainsExpectedValueForGivenHeader method declaration, so that they match the order they are stated in the method name. Updated all uses of method.

Renamed CreateCase_SuccessPage to CreateCaseSuccessPage just because I don't like underscores in file names.

Added stepdef methods for getting a new case (creating, loading and claiming the case) and for getting a case at a stage (creating, progressing and claiming at the designated stage), which enabled me to reduce the verbosity of background steps for most DCU scenarios.

Updated the step "I create a single {string} case and return to the dashboard" to use createNewCase stepdef method instead of createCase.createCSCaseOfType, enabling it to be used for any DECS casetype.

Removed the scenarios and steps that enabled the creation of a case without a document, as this covers no extra functionality then the scenarios that create a case with a document, and just extended the runtime of the regressions.

Updated the create case scenario that checks the stage a new case is at to also check the document type a document added at creation is given, extending the coverage of the regression.

Created the method moveCaseWithSpecifiedValuesToTriageStage, to reduce the number of methods we had for completing MPAM Triage stage with certain values specified. This should enable easier maintenance of this stage. The stepdef level still allows for variety in BDD language of what we are specifying.

Encapsulated some button pressed in methods unique to stage specific pages, so that unique waits could be included, or case data could be captured for latter assertions.

Reworded "I upload a 'second draft' document" to "I upload a 'replacement draft' document", as I felt this was clearer in its representation of the action undertaken by the user.

Created iGetAMINCaseAtTheDispatchStageThatShouldBeCopiedToNumber method, and workflow scenario for MIN cases at Dispatch stage that would go to Copy To Number 10 stage. Removed the similar scenario from the DCUEndToEnd feature.

Created feature and scenario for cases that are at Copy To Number 10 stage.

Amended references of UKVI to MPAM for all MPAM/MTS workflow related pages, methods and tags.

Replaced the UKVI_USER with an MPAM_USER, to better fit with the naming convention of most MPAM features/files.

Reduced the reference of page variables directly in stepdef files, and replaced with page methods.

Moved some summary tab and timeline tab scoped methods from workstacks file to the appropriate files.

Deleted out of date and unused fetchExistingDCUCases file. We have no plans to go back to using this method of getting a case at a workflow stage, and it has not been maintained with new improvements to DECS workflows.