Renamed EndToEndStepDefs.java to ProgressCaseStepDefs.java. Refactored logic from EndToEndStepDefs.java into separate 'ProgressCase' files. Amended all WCS scenarios to use the "iCreateACaseAndMoveItToAStage" step like all of the other workflows. Amended lots of logic for the new approach to progress case. Seperated the logic for SMC and IEDET from COMP/COMP2. Added more complaintType checks and conditional logic to COMP stage completion methods so they can be called for any complaint type, allowing for the deletion of a lot of duplicate methods and switch cases.
Each progress case file has a method that moves a case from a 'currentStage' to a 'targetStage'. It first gets the 'precedingStage' of the 'targetStage', then checks if the 'currentStage' matches the 'precedingStage'. If they dont match it calls itself with the 'precedingStage' passed in as the 'targetStage'. This allows the method to work back through all of the 'precedingStages', find the 'currentStage', and work the case forward to the 'targetStage'. It then completes the 'currentStage'.
If a none expected string is passed in as the currentStage, such as 'N/A', eventually the method works back through the 'precedingStages' till it returns 'Create New Case' as the 'precedingStage'. This escapes the looping method call and just creates a new case, allowing the method to start working back through the previous calls to the target stage.
Having this logic in files instead of the glue should mean we can make easier use of it in scenarios where we need to get a case to a stage, do something, then move it to a later stage for an assertion. At the minute this is quite a tedious thing to stepdef, with lots of duplication.